JSON Parser and Decompressor work together now.

+++ src/decompress_files.py
@@ -5,29 +5,43 @@ Author: Trinity Stroud
 
 import bz2
 import os, os.path
+import sys
+from shutil import copy
 
 def decompress():
     '''
     Decompresses all .json.bz2 files into readable json files
 
     Vars:
-        tweet_loc:  Location of twitter data
+        None
     Returns:
         None
     '''
     
     # Path to where GitHub repo is stored
-    userpath = r"C:\Users\[username]\Documents\GitHub"
+    userpath = os.getcwd()
     # Path to Twitter stream data from March 1, 2020
-    olddirpath = userpath + r"\CSC-426\twitter_stream_2020_03_01\03\01"
+    olddirpath = os.path.join(userpath, "twitter_stream_2020_03_01", "03", "01")
     # Path to where decompressed files will be stored
-    newdirpath = userpath + r"\CSC-426\resources\twitter_stream_2020_03_01"
+    newdirpath = os.path.join(userpath, "resources", "twitter_stream_2020_03_01")
+    # Creates new directory if it doesn't exist
+    if not os.path.exists(newdirpath):
+        os.mkdir(newdirpath)
 
     # List of directories in the current directory (represents hours)
     dirs = [name for name in os.listdir(olddirpath) if not os.path.isfile(name)]
 
+    size = len(dirs)
+    sys.stdout.write(f"[ ] Decompressing files... ")
+    sys.stdout.flush()
+
     # For each directory...
-    for dirname in dirs:
+    for i, dirname in enumerate(dirs):
+
+        s = f"%.3F" % (i/float(size) * 100) + "%"
+        sys.stdout.write(s)
+        sys.stdout.flush()
+
         path = os.path.join(olddirpath, dirname)
         # List of files in the current subdirectory
         files = [name for name in os.listdir(path)]
@@ -35,7 +49,19 @@ def decompress():
         # For each file...
         for filename in files:
             oldfilepath = os.path.join(olddirpath, dirname, filename)
-            newfilepath = os.path.join(newdirpath, dirname, filename[0:-4])
+            newfilepath = os.path.join(newdirpath, dirname)
+
+            # Creates new directory if it doesn't exist
+            if not os.path.exists(newfilepath):
+                os.mkdir(newfilepath)
+            # If file is already decompressed, just copies it over
+            if not oldfilepath.endswith(".bz2"):
+                copy(oldfilepath, newfilepath)
+                continue
+            newfilepath = os.path.join(newfilepath, filename[:-4])
+            # If decompressed file already present in resources
+            if os.path.exists(newfilepath):
+                continue
 
             newfile = open(newfilepath, 'wb')
             file = bz2.BZ2File(oldfilepath, 'rb')
@@ -47,3 +73,9 @@ def decompress():
             # Close the open files
             file.close()
             newfile.close()
+
+        sys.stdout.write("\b" * len(s))
+
+    sys.stdout.write("\r[+] Decompression Complete 100.000%\n")
+    sys.stdout.flush()
+
